Ménage

+++ src/Subsystems/Conveyor.cpp
@@ -4,27 +4,25 @@
 #include <SensorsState.hpp>
 #include "Iteration_time.hpp"
 
-
-
 namespace p28 {
 
 struct Step{
-	float clawServo;
-	float armServo;
+	int clawServo;
+	int armServo;
 	int conveyorSteps;
 	unsigned long stepTime;
 };
 const int Nsteps = 8;
 const int CimbStepInd = 5; // Step at which the stepper starts, keep this linked
 Step const sequence[Nsteps] = {
-	Step { .clawServo = kClaw_openAngle, .armServo = kArm_upAngle, .conveyorSteps = 0, .stepTime = 200},
-	Step { .clawServo = kClaw_openAngle, .armServo = kArm_downAngle, .conveyorSteps = 0, .stepTime = 1000},
-	Step { .clawServo = kClaw_closeAngle, .armServo = kArm_downAngle, .conveyorSteps = 0, .stepTime = 600},
-	Step { .clawServo = kClaw_closeAngle, .armServo = kArm_upAngle, .conveyorSteps = 0, .stepTime = 600},
-	Step { .clawServo = kClaw_openAngle, .armServo = kArm_upAngle, .conveyorSteps = 0, .stepTime = 50},
-	Step { .clawServo = kClaw_openAngle, .armServo = kArm_upAngle, .conveyorSteps = 0, .stepTime = 250},
-	Step { .clawServo = kClaw_openAngle, .armServo = kArm_upAngle, .conveyorSteps = kConveyor_stepsUntilUp, .stepTime = 11600}, // #define delay 
-	Step { .clawServo = kClaw_openAngle, .armServo = kArm_upAngle, .conveyorSteps = 0, .stepTime = 10000} // #define delay
+	Step { .clawServo = kClaw_openAngle, 	.armServo = kArm_upAngle, 	.conveyorSteps = 0, 						.stepTime = 200},
+	Step { .clawServo = kClaw_openAngle, 	.armServo = kArm_downAngle, .conveyorSteps = 0, 						.stepTime = 1000},
+	Step { .clawServo = kClaw_closeAngle, 	.armServo = kArm_downAngle, .conveyorSteps = 0, 						.stepTime = 600},
+	Step { .clawServo = kClaw_closeAngle, 	.armServo = kArm_upAngle, 	.conveyorSteps = 0, 						.stepTime = 600},
+	Step { .clawServo = kClaw_openAngle, 	.armServo = kArm_upAngle, 	.conveyorSteps = 0, 						.stepTime = 50},
+	Step { .clawServo = kClaw_openAngle, 	.armServo = kArm_upAngle, 	.conveyorSteps = 0, 						.stepTime = 250}, // Give the robot a chance to adjust itself (block still in claw zone)
+	Step { .clawServo = kClaw_openAngle, 	.armServo = kArm_upAngle, 	.conveyorSteps = kConveyor_stepsUntilUp,	.stepTime = 11600}, // #define delay 
+	Step { .clawServo = kClaw_openAngle, 	.armServo = kArm_upAngle,	.conveyorSteps = 0, 						.stepTime = 10000} // #define delay
 };
 
 void Conveyor::init()
@@ -44,7 +42,6 @@ void Conveyor::update(Iteration_time it_time)
 			startStepTime = it_time.time_ms;
 		}
 	}
-	
 }
 HardwareState Conveyor::aggregate(HardwareState hrdwState)
 {
@@ -61,8 +58,6 @@ bool Conveyor::over() const
 }
 bool Conveyor::just_dropped() const
 {
-	// Serial.print("Seq ind ");
-	// Serial.println(sequenceIndex);
 	if(sequenceIndex == CimbStepInd)
 		return true;
 	return false;
